# docs/05_API/OpenAPI.yaml
...
openapi: 3.1.0
info:
  title: MedRail API
  version: 0.1.0
  summary: Patient-consented health data layer under x402-paid intelligence endpoints, on Algorand.
  description: |-
    MedRail exposes **8 HTTP routes**: 3 priced with the x402 protocol (HTTP 402 +
    USDC settlement on Algorand) and 5 free. It also fronts an on-chain consent
    registry and audit log — the `MedRailConsent` smart contract, **App ID 768743428**
    on Algorand TestNet.

    ## What this API is not

    - **There is no database.** Durable state is Algorand global state, three box maps
      on App 768743428, and two static files compiled into the image.
    - **There is no LLM, model, or inference call.** The two "AI" endpoints are
      deterministic rule engines: 11 keyword rules with fixed weights, and a 14-row
      interaction table lookup. Every response carries a non-diagnostic disclaimer.
    - **There is no authentication.** No API keys, no tokens, no sessions, no accounts.
      Payment is the only gate, and it authenticates a *payment*, not a *caller*.
    - **There is no rate limiting** (requirement SEC-013 is NOT IMPLEMENTED).
    - **There is no idempotency.** No idempotency key, no deduplication, no replay
      window. A repeated paid call is a repeated payment.

    ## Middleware ordering — expect 402 before 400

    The x402 payment middleware is registered before every route handler
    (`api/src/app.ts:37-50`), so request-body validation never runs on an unpaid
...

...

...

...

...
  contact:
...
servers:
...
paths:
...
  /v1/consent/status:
    get:
      operationId: getConsentStatus
      tags: [consent]
      summary: Live on-chain consent validity for a (patient, requester, scope) triple (free)
      description: |-
        Free, unpaid, read-only. Executes the contract's `readonly` `check_access` method
        through `AtomicTransactionComposer.simulate()` — **zero fee, nothing submitted to
        the network** (requirement SEC-009).

        `granted` is `true` if and only if all three hold:

        1. the grant box exists;
        2. its `status` is 1 (`STATUS_GRANTED`);
        3. `expires_at == 0` (the "never expires" sentinel) **or**
           `Global.latest_timestamp < expires_at` (strict comparison, so a grant is already
           invalid *at* its expiry second).

        ## Two things that surprise integrators

        - **`scope` is free-form and is not normalised.** No trim, no case folding. The
          scope is hashed into the box key, so `"records:summary "` with a trailing space is
          a completely different grant and silently returns `granted: false`
          (requirement DATA-003).
        - **This free, unauthenticated endpoint requires the operator private key.** The
          simulated call needs a sender and a signer, so `OPERATOR_MNEMONIC` must be set
          (`api/src/services/algorand.ts:83-84`). Without it the endpoint returns 500.

        Two sequential algod round trips per request, with **no rate limit, no timeout, and
        no retry** — usable both to exhaust this service and to amplify traffic at AlgoNode
        (requirements SEC-013 and REL-003, both NOT IMPLEMENTED).
      x-medrail-status: UNVALIDATED
      x-medrail-requirements: [FR-013, SEC-009, SEC-010, SEC-013, DATA-003, REL-003, PERF-002]
      x-medrail-evidence: |-
        No automated test. Manually exercised by the reviewer: HTTP 200 in 505 ms, a single
        cold observation on a developer laptop. Not a p50, not an SLO; no latency budget
        exists.
      x-medrail-defects: [R-3]
      x-medrail-sideEffects: none
      x-medrail-idempotent: true
      x-medrail-rateLimited: false
      security: []
      parameters:
        - name: patient
          in: query
          required: true
          description: The patient's Algorand address. Exactly 58 characters; checksum NOT validated.
          schema:
            $ref: '#/components/schemas/AlgorandAddress'
          example: S56WIB3XLUOXSUGAI6MDJ3GN3RV2NSE7RVV7DGNZJKDCJR4OJLLA6H7HEE
        - name: requester
          in: query
          required: true
          description: The requester's Algorand address. Exactly 58 characters; checksum NOT validated.
          schema:
            $ref: '#/components/schemas/AlgorandAddress'
          example: CVYBERM3GTWGASZ7VUJLCBTVQA7AOPPTUW2Q3EISGR3XCRIKLM6SBIGUUA
        - name: scope
          in: query
          required: true
          description: |-
            Free-form consent scope. Minimum 1 character, no maximum, no normalisation.
            The API's own record route uses `records:summary`.
          schema:
            type: string
            minLength: 1
          example: 'records:summary'
      responses:
        '200':
          description: The live on-chain answer.
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/AccessControlAllowOrigin'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentStatusResponse'
              examples:
                revoked:
                  summary: The live triple on App 768743428 — box value decodes to status 2
                  value:
                    patient: S56WIB3XLUOXSUGAI6MDJ3GN3RV2NSE7RVV7DGNZJKDCJR4OJLLA6H7HEE
                    requester: CVYBERM3GTWGASZ7VUJLCBTVQA7AOPPTUW2Q3EISGR3XCRIKLM6SBIGUUA
                    scope: 'records:summary'
                    granted: false
        '400':
          description: |-
            One or more query parameters failed validation. Note the route-specific message.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
              example:
                error: 'invalid query — expected ?patient=&requester=&scope='
                details:
                  formErrors: []
                  fieldErrors:
                    patient: ['String must contain exactly 58 character(s)']
                    scope: ['String must contain at least 1 character(s)']
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/consent/arc56:
    get:
      operationId: getConsentArc56
      tags: [consent]
      summary: The compiled ARC-56 application specification (free)
      description: |-
        Serves `contracts/artifacts/MedRailConsent.arc56.json` verbatim, so any third party
        can construct their own ABI calls against App 768743428 without this repository.

        Top-level keys: `name`, `structs` (`GrantRecord`, `AuditEntry`), `methods` (13),
        `arcs` (`[22, 28]`), `networks` (**an empty object — the deployed App ID is not
        recorded in the spec; use `/v1/consent/app-info`**), `state`, `bareActions`,
        `sourceInfo`.

        **Implementation note:** the file is `existsSync`-ed, `readFileSync`-read and
        `JSON.parse`-d on **every request** (`api/src/app.ts:63-68`) — a synchronous ~54 KB
        disk read and parse in the event loop, on a free, unauthenticated, unrate-limited
        route. Reading it once at module load would be strictly better. RECOMMENDED.
      x-medrail-status: UNVALIDATED
      x-medrail-requirements: [FR-015]
      x-medrail-evidence: No automated test. Verified by reading api/src/app.ts:63-69.
      x-medrail-sideEffects: none
      x-medrail-idempotent: true
      x-medrail-rateLimited: false
      security: []
      responses:
        '200':
          description: The ARC-56 application specification.
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/AccessControlAllowOrigin'
          content:
            application/json:
              schema:
                type: object
                description: |-
                  An ARC-56 Contract object. Not modelled field by field here; the served
                  document is authoritative.
                additionalProperties: true
                required: [name, structs, methods, arcs, state, bareActions]
                properties:
                  name:
                    type: string
                    const: MedRailConsent
                  arcs:
                    type: array
                    items:
                      type: integer
                    examples:
                      - [22, 28]
        '404':
          description: |-
            The compiled artifact is not present on disk at
            `__dirname/../../contracts/artifacts/MedRailConsent.arc56.json`. In a container
            this means the Dockerfile COPY step was skipped or the path resolution differs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                error: 'ARC-56 spec not found — has the contract been compiled?'
...
components:
  securitySchemes:
    x402Payment:
      type: apiKey
      in: header
      name: PAYMENT-SIGNATURE
      description: |-
        **Not an authentication scheme.** Modelled as `apiKey` only because OpenAPI has no
        native x402 type. The `PAYMENT-SIGNATURE` header carries a base64-encoded JSON
        `PaymentPayload` containing a signed Algorand transaction group, produced by an
        x402 v2 client such as `@x402/fetch` with `@x402/avm`'s `ExactAvmScheme`.

        It proves that **a payment settled**. It does **not** identify the caller to the
        route handler, and MedRail's handlers never decode it — which is precisely defect
        S-1. Every operation in this document therefore declares `security: []`: there is no
        authentication anywhere in this API.
...
  headers:
...
    AccessControlAllowOrigin:
      description: Always `*` — CORS is fully open by design (`api/src/app.ts:23`, requirement NFR-006).
      required: false
      schema:
        type: string
        const: '*'
...
  schemas:
    AlgorandAddress:
      type: string
      minLength: 58
      maxLength: 58
      description: |-
        A base32 Algorand address.

        **Validation is length only** — the zod schema is `z.string().length(58)`. The
        checksum is NOT validated (requirement SEC-010, NOT IMPLEMENTED). A 58-character
        string that is not a valid address reaches `algosdk.decodeAddress` and throws,
        producing **HTTP 500** with the internal message echoed to the caller (defect R-3).
        The `pattern` below documents the true character set; the implementation does not
        enforce it.
      pattern: '^[A-Z2-7]{58}$'
      examples:
        - S56WIB3XLUOXSUGAI6MDJ3GN3RV2NSE7RVV7DGNZJKDCJR4OJLLA6H7HEE
        - CVYBERM3GTWGASZ7VUJLCBTVQA7AOPPTUW2Q3EISGR3XCRIKLM6SBIGUUA
...
    ConsentStatusResponse:
      type: object
      required: [patient, requester, scope, granted]
      properties:
        patient:
          $ref: '#/components/schemas/AlgorandAddress'
        requester:
          $ref: '#/components/schemas/AlgorandAddress'
        scope:
          type: string
        granted:
          type: boolean
          description: |-
            `true` iff the grant box exists, `status == 1`, and either `expires_at == 0`
            (never expires) or the consensus timestamp is strictly less than `expires_at`.
...
    ErrorBody:
      type: object
      required: [error]
      additionalProperties: true
      description: |-
        The universal error envelope. **There is no machine-readable error code anywhere in
        this API** — clients must match on prose, which is brittle. A stable `code` field is
        RECOMMENDED; see `./API_Error_Catalog.md`.
      properties:
        error:
          type: string
          description: |-
            A human-readable message. On a 500 this is `err.message` echoed verbatim from
            the caught exception (`api/src/app.ts:60`), which leaks internal detail to
            unauthenticated callers (requirement SEC-011, NOT IMPLEMENTED; defect R-3).
    ValidationErrorBody:
      allOf:
        - $ref: '#/components/schemas/ErrorBody'
        - type: object
          required: [details]
          properties:
            details:
              $ref: '#/components/schemas/ZodFlattenedError'
    ZodFlattenedError:
      type: object
      description: |-
        The output of zod's `.flatten()` (zod resolves to 3.25.76 under the declared
        `^3.24.1`). Verified empirically against the installed package.
      required: [formErrors, fieldErrors]
      properties:
        formErrors:
          type: array
          description: |-
            Object-level issues. **Empty for every schema in this codebase**, because none
            uses an object-level `.refine()`.
          items:
            type: string
        fieldErrors:
          type: object
          description: Per-field messages, keyed by field name.
          additionalProperties:
            type: array
            items:
              type: string
          examples:
            - symptoms: [Required]
            - medications: ['Array must contain at least 2 element(s)']
            - patientId: ['String must contain exactly 58 character(s)']
  responses:
...
    InternalError:
      description: |-
        An unhandled exception reached `app.onError` (`api/src/app.ts:58-61`), which logs it
        with `console.error` and returns **`err.message` verbatim** to the caller
        (requirement SEC-011, NOT IMPLEMENTED).

        Known causes, all reproduced or verified:

        - **R-1** — the facilitator is unreachable, so the 402 cannot be built:
          `"Failed to initialize: no supported payment kinds loaded from any facilitator."`
          Note the client gets **no `PAYMENT-REQUIRED` header, no `Retry-After`, and not a
          503** (requirement REL-001, NOT IMPLEMENTED).
        - **R-3** — a 58-character string that is not a valid address:
          `"wrong checksum for address"`, or `"Invalid base32 characters"` for a
          58-character string outside the base32 alphabet. A client input error surfaced as
          a server error.
        - **D-1** — `CONSENT_APP_ID` is unset and `contracts/artifacts/deploy_testnet.json`
          is absent, which is the default inside the container image:
          `"CONSENT_APP_ID is not set and contracts/artifacts/deploy_testnet.json was not found. Deploy the contract first (see docs/DEPLOYMENT.md)."`
        - `OPERATOR_MNEMONIC` unset: `"OPERATOR_MNEMONIC is not set — see docs/DEPLOYMENT.md"`.
        - **R-4** — algod unreachable or slow. `new algosdk.Algodv2("", server, "")` has no
          timeout and no retry (requirement REL-003, NOT IMPLEMENTED).

        Full catalogue: `./API_Error_Catalog.md`.
      headers:
        Access-Control-Allow-Origin:
          $ref: '#/components/headers/AccessControlAllowOrigin'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          examples:
            facilitatorDown:
              summary: R-1 — facilitator unreachable
              value:
                error: 'Failed to initialize: no supported payment kinds loaded from any facilitator.'
            badChecksum:
              summary: R-3 — 58 characters, invalid checksum
              value:
                error: wrong checksum for address
            appIdMissing:
              summary: D-1 — no App ID configured, the container default
              value:
                error: 'CONSENT_APP_ID is not set and contracts/artifacts/deploy_testnet.json was not found. Deploy the contract first (see docs/DEPLOYMENT.md).'
...